more file / func organization
