+++ lib/modder/module_deps.go
@@ -1,0 +1,1 @@
+package modder
